sdks/cosmosdb/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cosmosdb"
version = "0.1.10"
authors = [
    {name="Marcel Claramunt", email="marcel@moveread.com"}
]
description = "Python SDK for Azure Cosmos DB NoSQL"
dependencies = [
    "azure-cosmos", "aiohttp", "haskellian", "pydantic"
]
requires-python = ">=3.10"
readme = {file="README.md", content-type="text/markdown"}


[project.urls]
home = "https://github.com/moveread/azure-sdks/tree/main/sdks/cosmosdb"
repo = "https://github.com/moveread/azure-sdks"

# [project.optional-dependencies]
# test = [
#   "pytest < 5.0.0",
#   "pytest-cov[all]"
# ]